resolves 188, resolves 191

+++ logic/src/main/java/a75f/io/logic/bo/building/vav/VavProfileConfiguration.java
@@ -25,5 +25,5 @@ public class VavProfileConfiguration extends BaseProfileConfiguration
     public int minDamperHeating;
     public int maxDamperHeating;
     
-    public int temperaturOffset;
+    public double temperaturOffset;
 }
